+++ CumulusMX/Cumulus.cs
@@ -2828,6 +2828,7 @@ namespace CumulusMX
 
 						Interlocked.Increment(ref taskCount);
 
+						// Create a fixed index reference
 						var idx = i;
 						tasklist.Add(Task.Run(async () =>
 						{
@@ -2932,6 +2933,7 @@ namespace CumulusMX
 					}
 
 					var remotefile = GetRemoteFileName(ActiveExtraFiles[i].DestFilename, DateTime.Now);
+					// Create a fixed index reference
 					var idx = i;
 
 					Interlocked.Increment(ref taskCount);
@@ -2947,10 +2949,10 @@ namespace CumulusMX
 
 							// all checks OK, file needs to be uploaded
 							// Is this an incremental log file upload?
-							if (ActiveExtraFiles[i].Incremental && !ActiveExtraFiles[i].Binary)
+							if (ActiveExtraFiles[idx].Incremental && !ActiveExtraFiles[idx].Binary)
 							{
 								LogDebugMessage($"Realtime[{cycle}]: Uploading extra web incremental file {uploadfile} to {remotefile} ({(incremental ? $"Incremental - {linesAdded} lines" : "Full file")})");
-								if (await UploadString(phpUploadHttpClient, incremental, string.Empty, data, remotefile, cycle, ActiveExtraFiles[i].Binary, ActiveExtraFiles[i].Utf8, true, ActiveExtraFiles[i].logFileLastLineNumber))
+								if (await UploadString(phpUploadHttpClient, incremental, string.Empty, data, remotefile, cycle, ActiveExtraFiles[idx].Binary, ActiveExtraFiles[idx].Utf8, true, ActiveExtraFiles[idx].logFileLastLineNumber))
 								{
 									ActiveExtraFiles[idx].logFileLastLineNumber += linesAdded;
 								}
@@ -2959,16 +2961,16 @@ namespace CumulusMX
 							{
 								LogDebugMessage($"Realtime[{cycle}]: Uploading extra web file {uploadfile} to {remotefile}");
 
-								if (ActiveExtraFiles[i].Process)
+								if (ActiveExtraFiles[idx].Process)
 								{
 									LogDebugMessage($"Realtime[{cycle}]: Processing extra web file {uploadfile}");
-									var str = await ProcessTemplateFile2StringAsync(uploadfile, false, ActiveExtraFiles[i].Utf8);
+									var str = await ProcessTemplateFile2StringAsync(uploadfile, false, ActiveExtraFiles[idx].Utf8);
 
-									_ = await UploadString(phpUploadHttpClient, false, string.Empty, str, remotefile, cycle, ActiveExtraFiles[i].Binary, ActiveExtraFiles[i].Utf8);
+									_ = await UploadString(phpUploadHttpClient, false, string.Empty, str, remotefile, cycle, ActiveExtraFiles[idx].Binary, ActiveExtraFiles[idx].Utf8);
 								}
 								else
 								{
-									_ = await UploadFile(phpUploadHttpClient, uploadfile, remotefile, cycle, ActiveExtraFiles[i].Binary, ActiveExtraFiles[i].Utf8);
+									_ = await UploadFile(phpUploadHttpClient, uploadfile, remotefile, cycle, ActiveExtraFiles[idx].Binary, ActiveExtraFiles[idx].Utf8);
 								}
 							}
 						}
@@ -4800,6 +4802,7 @@ namespace CumulusMX
 
 				var item = (new Settings.ExtaWebFilesItem()
 				{
+					Enabled = ini.GetValue("FTP site", "ExtraEnable" + i, false),
 					LocalFilename = ini.GetValue("FTP site", "ExtraLocal" + i, string.Empty),
 					DestFilename = ini.GetValue("FTP site", "ExtraRemote" + i, string.Empty),
 					Process = ini.GetValue("FTP site", "ExtraProcess" + i, false),
@@ -4812,14 +4815,6 @@ namespace CumulusMX
 					Incremental = ini.GetValue("FTP site", "ExtraIncLogFile" + i, false)
 				});
 
-				if (ini.ValueExists("FTP site", "ExtraEnable" + i))
-				{
-					item.Enabled = ini.GetValue("FTP site", "ExtraEnable" + i, false);
-				}
-				else
-				{
-					item.Enabled = !string.IsNullOrEmpty(item.LocalFilename) && !string.IsNullOrEmpty(item.DestFilename);
-				}
 
 				if (item.Binary)
 				{